# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "DigiSchool API",
        "description": "API REST pour la gestion d'ecoles DigiSchool.\n\n## Authentification\nLa plupart des endpoints necessitent un token JWT.\n1. Utilisez `POST /api/auth/login` pour obtenir un token\n2. Cliquez sur le bouton \"Authorize\" et entrez: `Bearer <votre_token>`\n\n## Multi-tenancy\nLe systeme est multi-tenant base sur l'ecole (ecoleId).\nL'ecoleId est automatiquement extrait du token JWT.\n\n## Comptes de test\n| Role | Email | Telephone | Password |\n|------|-------|-----------|----------|\n| Admin | admin@digischool.cm | +237600000000 | Admin@2025 |\n| Directeur | smbarga@lavictoire.cm | +237677123456 | Directeur@2025 |\n| Enseignant | jpkamga@lavictoire.cm | +237670111222 | Enseignant@2025 |\n| Secretaire | catangana@lavictoire.cm | +237660555666 | Secretaire@2025 |\n| Parent | fnkoulou@gmail.com | +237691666777 | Parent@2025 |\n",
        "contact": {
            "name": "DigiSchool Support",
            "url": "https://digischool.cm",
            "email": "support@digischool.cm"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://digischool.cm/license"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Serveur de developpement local"
        },
        {
            "url": "https://api.digischool.cm",
...
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/api/classes/{id}": {
            "get": {
                "tags": [
                    "classe-controller"
                ],
                "operationId": "getById_7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClasseDto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "classe-controller"
                ],
                "operationId": "update_7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClasseDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClasseDto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "classe-controller"
                ],
                "operationId": "delete_7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/arrondissements/{id}": {
            "get": {
                "tags": [
                    "arrondissement-controller"
                ],
                "operationId": "getById_8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArrondissementDto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "arrondissement-controller"
                ],
                "operationId": "update_8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ArrondissementDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArrondissementDto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "arrondissement-controller"
                ],
                "operationId": "delete_8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/classes": {
            "get": {
                "tags": [
                    "classe-controller"
                ],
                "operationId": "getAll_7",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ClasseDto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "classe-controller"
                ],
                "operationId": "create_7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClasseDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClasseDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/arrondissements": {
            "get": {
                "tags": [
                    "arrondissement-controller"
                ],
                "operationId": "getAll_8",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ArrondissementDto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "arrondissement-controller"
                ],
                "operationId": "create_8",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ArrondissementDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArrondissementDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/villes/arrondissement/{arrondissementId}": {
            "get": {
                "tags": [
                    "ville-controller"
                ],
                "operationId": "getByArrondissementId",
                "parameters": [
                    {
                        "name": "arrondissementId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/VilleDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/classes/ecole/{ecoleId}": {
            "get": {
                "tags": [
                    "classe-controller"
                ],
                "operationId": "getByEcoleId",
                "parameters": [
                    {
                        "name": "ecoleId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ClasseDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/arrondissements/departement/{departementId}": {
            "get": {
                "tags": [
                    "arrondissement-controller"
                ],
                "operationId": "getByDepartementId",
                "parameters": [
                    {
                        "name": "departementId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ArrondissementDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "VilleDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "code": {
                        "type": "string"
                    },
                    "nom": {
                        "type": "string"
                    },
                    "codePostal": {
                        "type": "string"
                    },
                    "arrondissementId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "arrondissementNom": {
                        "type": "string"
                    }
                }
            },
...
            "ClasseDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nomClasse": {
                        "type": "string"
                    },
                    "niveau": {
                        "type": "string",
                        "enum": [
                            "MATERNELLE",
                            "PRIMAIRE",
                            "COLLEGE",
                            "LYCEE",
                            "NURSERY",
                            "PRIMARY",
                            "SECONDARY",
                            "HIGH_SCHOOL"
                        ]
                    },
                    "sousSysteme": {
                        "type": "string",
                        "enum": [
                            "FRANCOPHONE",
                            "ANGLOPHONE",
                            "BILINGUE"
                        ]
                    },
                    "section": {
                        "type": "string"
                    },
                    "capacite": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "statut": {
                        "type": "string",
                        "enum": [
                            "ACTIVE",
                            "INACTIVE",
                            "ARCHIVEE"
                        ]
                    },
                    "effectifActuel": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "fraisScolarite": {
                        "type": "number",
                        "format": "double"
                    },
                    "description": {
                        "type": "string"
                    },
                    "ecoleId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "ecoleNom": {
                        "type": "string"
                    },
                    "anneeScolaireId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "anneeScolaireLibelle": {
                        "type": "string"
                    },
                    "titulaireId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "titulaireNom": {
                        "type": "string"
                    }
                }
            },
            "ArrondissementDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "code": {
                        "type": "string"
                    },
                    "nom": {
                        "type": "string"
                    },
                    "departementId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "departementNom": {
                        "type": "string"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT obtenu via POST /api/auth/login. Entrez: Bearer <token>",
                "name": "bearerAuth",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}